Hash properly
