+++ coffea/util.py
@@ -2,6 +2,7 @@
 
 """
 import awkward
+import hashlib
 
 from awkward.util import numpy
 import numba
@@ -29,3 +30,23 @@ def save(output, filename):
     '''
     with lz4.frame.open(filename, 'wb') as fout:
         cloudpickle.dump(output, fout)
+
+
+def _hex(string):
+    try:
+        return string.hex()
+    except AttributeError:
+        return "".join("{:02x}".format(ord(c)) for c in string)
+
+
+def _ascii(maybebytes):
+    try:
+        return maybebytes.decode('ascii')
+    except AttributeError:
+        return maybebytes
+
+
+def _hash(items):
+    # python 3.3 salts hash(), we want it to persist across processes
+    x = hashlib.md5(bytes(';'.join(str(x) for x in items), 'ascii'))
+    return int(x.hexdigest()[:16], base=16)
